pyproject.toml
[tool.poetry]
name = "sunshine-input-bridge"
version = "0.1.0"
description = ""
authors = ["Dennis George <dennispg@gmail.com>"]
readme = "README.md"
packages = [{include = "sunshine_input_bridge"}]

[tool.poetry.dependencies]
python = "^3.9"
evdev = "^1.6.1"
asyncio = "^3.4.3"
pyudev = "^0.24.1"
python-uinput = "^0.11.2"
python-pidfile = "^3.1.1"
argparse = "^1.4.0"
pydantic = "^1.10.9"
pyyaml = "^6.0"
requests-unixsocket = "^0.3.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^23.3.0"
pytest = "^7.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
